Move responsibility to delay processing of new files detected from InstantUploadsObservers into InstantUploadsHandler
